GetTranslationEntryPairs moved to keys
